Connect database to updating a journal entry on the form
