+++ src/pages/DashboardPage/DashboardPage.tsx
@@ -11,7 +11,7 @@ import { useJournalEntries } from '../../api/journalEntries/journalEntries';
 
 import './DashboardPage.css';
 
-export interface mockData {
+export interface JournalEntry {
     id: number;
     authorId: string;
     date: string;
@@ -24,114 +24,20 @@ export interface mockData {
     probiotics: boolean;
 }
 
-export const mockEntries: mockData[]= [
-    {
-        "id": 1,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-07T19:58:57.000Z",
-        "waterIntake": 1,
-        "proteinIntake": 1,
-        "exercise": 1,
-        "kegels": 1,
-        "garlandPose": 1,
-        "prenatalVitamins": true,
-        "probiotics": true 
-    },
-    {
-        "id": 2,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-08T19:58:57.000Z",
-        "waterIntake": 2,
-        "proteinIntake": 2,
-        "exercise": 2,
-        "kegels": 2,
-        "garlandPose": 2,
-        "prenatalVitamins": true,
-        "probiotics": false
-     },
-     {
-        "id": 3,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-09T19:58:57.000Z",
-        "waterIntake": 3,
-        "proteinIntake": 3,
-        "exercise": 3,
-        "kegels": 3,
-        "garlandPose": 3,
-        "prenatalVitamins": false,
-        "probiotics": false
-     },
-     {
-        "id": 4,
-        "authorId": "DRYvbhpN9UPqJPcbKV21TH3Yp8N2",
-        "date": "2022-03-10T19:58:57.000Z",
-        "waterIntake": 4,
-        "proteinIntake": 4,
-        "exercise": 4,
-        "kegels": 4,
-        "garlandPose": 4,
-        "prenatalVitamins": true,
-        "probiotics": true 
-     },
-     {
-        "id": 5,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-09T19:58:57.000Z",
-        "waterIntake": 5,
-        "proteinIntake": 5,
-        "exercise": 5,
-        "kegels": 5,
-        "garlandPose": 5,
-        "prenatalVitamins": false,
-        "probiotics": false
-     },
-     {
-        "id": 6,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-09T19:58:57.000Z",
-        "waterIntake": 6,
-        "proteinIntake": 6,
-        "exercise": 6,
-        "kegels": 6,
-        "garlandPose": 6,
-        "prenatalVitamins": false,
-        "probiotics": false
-     },
-     {
-        "id": 7,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-09T19:58:57.000Z",
-        "waterIntake": 7,
-        "proteinIntake": 7,
-        "exercise": 7,
-        "kegels": 7,
-        "garlandPose": 7,
-        "prenatalVitamins": false,
-        "probiotics": false
-     },
-     {
-        "id": 8,
-        "authorId": "fV5De0bivMRqBoHxJuwT4UwFJtT2",
-        "date": "2022-03-09T19:58:57.000Z",
-        "waterIntake": 8,
-        "proteinIntake": 8,
-        "exercise": 8,
-        "kegels": 8,
-        "garlandPose": 8,
-        "prenatalVitamins": false,
-        "probiotics": false
-     },
-]
-
 const DashboardPage = () => {
     const user = useContext(AuthContext);
     const navigate = useNavigate();
-    const { data } = useJournalEntries(user.id);
+    const { data, isFetching } = useJournalEntries(user.id);
 
     const handleNavigateToJournalEntryForm = (callback: () => void) => {
         return () => {
             callback()
         }
+    }
+    if (isFetching) {
+        return (
+            <p>Fetching data...</p>
+        )
     }
 
     return user.isLoggedIn ? (
@@ -140,7 +46,7 @@ const DashboardPage = () => {
             {data.length ? 
             (
                 <Box className='dashboard-journal-entries-container'>
-                    {data.map((entry: mockData) => {
+                    {data.map((entry: JournalEntry) => {
                         return <JournalEntryCard entry={entry} key={entry.id}/>
                     })}
                 </Box>
